refactor: useQuickLink 훅에서 QuickLink 컴포넌트로 변경
